+++ model_tuning/performance.py
@@ -7,93 +7,15 @@ from sentence_transformers import SentenceTransformer
 from sentence_transformers import util
 from torch import Tensor
 
+from utils.parse_and_extract_loinc_names import parse_snoinc_extracts
+
 MODEL_NAME = "all-MiniLM-L6-v2"
-SNOINC_CODES_FILE = "../data/snoinc_extracts/loinc_lab_names_20250911.csv"
-# use date in filename to keep track of versions
+SNOINC_CODES_FILE = "../data/snoinc_extracts/loinc_lab_names_20251008.csv"
 DATE = SNOINC_CODES_FILE.split("_")[-1].split(".")[0]
 EMBEDDING_CACHE_DIR = "../data/training_files/embeddings/"
 EMBEDDING_FILE = f"loinc_lab_names_{MODEL_NAME.replace('/', '_')}_{DATE}"
 VALIDATION_FILE = "../data/training_files/validation_toy.txt"
 K_VALUES = [1, 3, 5, 10]
-
-
-def parse_snoinc_extracts(
-    extract_path: str,
-    short_name_col: int = 1,
-    long_name_col: int = 2,
-    display_name_col: int = 3,
-    skip_first: bool = True,
-):
-    """
-    Given a path to an extract file of information on various LOINC codes,
-    parse the rows of that file in to three discrete lists corresponding to
-    the long common names, short names, and display names of those codes.
-    The file is expected to be a pipe-delimited text file in which each
-    LOINC code is expected to represent a single line.
-
-    :param extract_path: The path to the extract file to parse.
-    :param short_name_col: The column of the pipe file containing the
-      short name for a given LOINC code.
-    :param long_name_col: The column of the pipe file containing the long
-      common name for a given LOINC code.
-    :param display_name_col: The column of the pipe file containing the
-      display name for a given LOINC code.
-    :param skip_first: Optionally, a boolean indicating whether to skip the
-      first line of the file, if it is a header row.
-    :returns: A tuple of three lists, one for eaech name variant.
-    """
-    long_common_names = []
-    short_names = []
-    display_names = []
-
-    with open(extract_path, "r", encoding="utf-8") as fp:
-        lines_seen = 0
-        for line in fp:
-            if lines_seen == 0:
-                lines_seen += 1
-                if skip_first:
-                    continue
-            if line.strip() != "":
-                names = line.strip().split("|")
-                # Skip lines that aren't real entries (formatting artifacts)
-                if len(names) >= 4:
-                    long_common_names.append(names[long_name_col].strip())
-                    short_names.append(names[short_name_col].strip())
-                    display_names.append(names[display_name_col].strip())
-
-    for name_list in [long_common_names, short_names, display_names]:
-        name_list = [x for x in name_list if not x == ""]
-
-    return long_common_names, short_names, display_names
-
-
-def embed_loinc_names(
-    model: SentenceTransformer, name_list: List[str], dest: str, save_embeddings: bool = False
-):
-    """
-    Use a SentenceTransformers model to embed the standard name codes for
-    a given set of LOINC values. These embeddings form the "Vector DB" that
-    will be used for semantic search on the examples-to-evaluate. Optionally,
-    save the embeddings to disk since computing them is time-consuming.
-
-    :param model: The Sentence Transformers model to use for embedding.
-    :param name_list: A list of strings to embed into the Vector DB.
-    :param save_embeddings: Optionally, a boolean in dicating whether to persist
-      the computed embeddings to disk.
-    :returns: The computed embeddings.
-    """
-    corpus_embeddings = model.encode(
-        name_list, batch_size=16, show_progress_bar=True, convert_to_tensor=True
-    )
-
-    # make sure just the directory exists
-    os.makedirs(EMBEDDING_CACHE_DIR, exist_ok=True)
-
-    if save_embeddings:
-        with open(EMBEDDING_CACHE_DIR + dest, "wb") as fp:
-            pickle.dump({"codes": name_list, "embeddings": corpus_embeddings}, fp)
-
-    return corpus_embeddings
 
 
 def predict_and_evaluate_validation_set(
@@ -176,20 +98,18 @@ if __name__ == "__main__":
             cache_data = pickle.load(fp)
             name_codes = cache_data["codes"]
             embeddings = cache_data["embeddings"]
+
+            print("Loading validation set...")
+            examples = []
+            with open(VALIDATION_FILE, "r") as fp:
+                for line in fp:
+                    if line.strip() != "":
+                        examples.append(line.split("|"))
+
+            print("Predicting and computing stats for validation set...")
+            for k in K_VALUES:
+                print(f"  Trial: Value for Top-K is {k}")
+                predict_and_evaluate_validation_set(model, embeddings, name_codes, examples, k)
+
     else:
-        print("No cache found, performing embedding.")
-        name_codes = lcns + sns + dns
-        print("  This might take a while...")
-        embeddings = embed_loinc_names(model, name_codes, EMBEDDING_FILE, save_embeddings=True)
-
-    print("Loading validation set...")
-    examples = []
-    with open(VALIDATION_FILE, "r") as fp:
-        for line in fp:
-            if line.strip() != "":
-                examples.append(line.split("|"))
-
-    print("Predicting and computing stats for validation set...")
-    for k in K_VALUES:
-        print(f"  Trial: Value for Top-K is {k}")
-        predict_and_evaluate_validation_set(model, embeddings, name_codes, examples, k)
+        print("No embeddings found, please run embedding.py to compute vectors first.")
